+++ package.json
@@ -50,23 +50,30 @@
         "@types/react": "*"
     },
     "dependencies": {
-        "@material-ui/core": "^4.11.3",
-        "@material-ui/icons": "^4.11.2",
+        "@material-ui/core": "^4.11.4",
         "embla-carousel": "^4.1.5",
+        "onyxia-ui": "^0.3.0",
         "powerhooks": "^0.0.37",
+        "react-code-blocks": "^0.0.8",
         "react-markdown": "^5.0.3",
         "react-svg": "^13.0.4",
+        "tsafe": "^0.4.0",
         "tss-react": "^0.0.12",
-        "typeface-work-sans": "^1.1.13",
-        "react-code-blocks": "^0.0.8",
-        "onyxia-ui": "^0.3.0",
-        "tsafe": "^0.4.0"
+        "typeface-work-sans": "^1.1.13"
     },
     "devDependencies": {
+        "@storybook/addon-actions": "^6.2.9",
+        "@storybook/addon-essentials": "^6.2.9",
+        "@storybook/addon-links": "^6.2.9",
+        "@storybook/node-logger": "^6.2.9",
+        "@storybook/preset-create-react-app": "^3.1.7",
+        "@storybook/react": "^6.2.9",
         "@types/node": "^15.0.2",
         "@types/react": "^17.0.4",
+        "@types/react-dom": "^17.0.0",
         "@typescript-eslint/eslint-plugin": "^4.15.1",
         "@typescript-eslint/parser": "^4.15.1",
+        "copyfiles": "^2.4.1",
         "eslint": "^7.20.0",
         "eslint-config-prettier": "^7.2.0",
         "evt": "^2.0.0-beta.18",
@@ -74,18 +81,10 @@
         "lint-staged": "^10.5.4",
         "prettier": "^2.2.1",
         "react": "^16.14.0",
+        "react-code-blocks": "^0.0.8",
         "react-dom": "^17.0.2",
         "react-scripts": "4.0.3",
         "react-svg": "^13.0.4",
-        "typescript": "4.2.2",
-        "copyfiles": "^2.4.1",
-        "@types/react-dom": "^17.0.0",
-        "react-code-blocks": "^0.0.8",
-        "@storybook/addon-actions": "^6.2.9",
-        "@storybook/addon-essentials": "^6.2.9",
-        "@storybook/addon-links": "^6.2.9",
-        "@storybook/node-logger": "^6.2.9",
-        "@storybook/preset-create-react-app": "^3.1.7",
-        "@storybook/react": "^6.2.9"
+        "typescript": "4.2.2"
     }
 }
