+++ src/ReviewSlider.tsx
@@ -1,8 +1,6 @@
 import { useEmblaCarousel } from "embla-carousel/react";
 import { createUseClassNames } from "./theme";
 import { useConstCallback } from "powerhooks";
-import ArrowBackIosIcon from "@material-ui/icons/ArrowBackIos";
-import ArrowForwardIosIcon from "@material-ui/icons/ArrowForwardIos";
 import Paper from "@material-ui/core/Paper";
 import Typography from "@material-ui/core/Typography";
 import ReactMarkdown from "react-markdown";
