+++ src/components/ThumbNails.tsx
