--- src/lib/index.ts
@@ -1,6 +1,0 @@
-export * from "./Gitlanding";
-/*import twitterLogoUrl from "./assets/svg/twitter.svg";
-import youtubeLogoUrl from "./assets/svg/youtube.svg";
-import redditLogoUrl from "./assets/svg/reddit.svg";
-
-export { twitterLogoUrl, youtubeLogoUrl, redditLogoUrl };*/
